requirements.txt
Django==1.9
Markdown==2.6.8
Pillow==3.3.0
django-filter==1.0.2
djangorestframework==3.6.2
freeze==1.0.10
six==1.10.0


